Features and bugfixes
Moved logging variables and instance to class variable in RobBotClient

Improved feature: Scheduled events are now executed on the second
that they are defined. Events are represented with the Event() dataclass.

The purge_runtime() method counts down the seconds to midnight which is
when the purge and various updates in the service will occur.