New module with Reminder() class. This class is designed to hold all instances of Event() objects and return an instance of Event() if and when the date and time for that event is current.
the get() method is designed to be inside an infinity loop with an
async method, using asyncio.sleep() to query this method every 1 second
for the entire lifespan of the program.

When an event is eventually found to carry current date and time, it
will be returned to the caller.