Improvements and refactoring
* Removed channel from env file
* Removed add_events() method due to Reminder() class implementation
* New logic for auto messages- now, looping over reminder.get()
indefinitely until an event is returned, and then displayed in chat.

* Reoccuring reminders such as the friday message or anything else,
is now added outside of the object logic before main loop is started.
This way, events are not hard-coded in the classes like before and
are easy to change upon demand.

The setup_reminders() method will create Event() objects for every
lesson found for a given day. These will then be added to the Reminder
object and thus, a reminder will sound in the discord channel.