This commit includes features and bugfixes.
Bug fix: Schedule chat return would not display days correctly
if one day has less or more than two events. (_get_schedule_phrase method)

Feature: The ability to now record any time the bot does not know how
to respond to a chat. These entries are saved in a JSON file, and
the chat messages are represented by the dataclass UnrecognizedCommand.
Upon bot boot, the json file is parsed to retrieve old data in to ram.
This will later be used for improving the bot and also implementing
machine learning.

Feature / Refactoring:
Due to the more intelligent Event() object, the need for string formatting
every time one of these instances are to be displayed is gone.
This resulted in cleaner code in the _get_remembered_events method.