This commit includes features.
Event() class refactored and now has the field 'alarm' which is a time
stamp representing the acutal time for the event, minus a timedelta.
A much easier way to keep the time of the event and the time to remind,
separate. Thus, this property is now fetched in the get() method.