New module: Scraper.
The Scraper class in the scraper module file, is used as a way
to scrape the lunch restaurant website at campus. The logic in this
object is restricted to this site only due to various methods of
extracting data.

The object returns a Menu() instance which holds a structured
set of data with mappings to each weekday, mapped by number of weekday.