New module: menu
The Menu() class in the menu module is designed to structure parsed
HTML data from BeautifulSoup, and return the menu for the current
day according to data from the website.

The menu on the site is structured as...

dayofweek
menuitem
menuitem
menuitem
menuitem

dayofweek
menuitem
menuitem

...

Hence, the iteration can look for these strings (dayofweek) and
increment the key counter, and append each menu item to the
list object on the value side of that key in a dictionary.
This happens in the serialize() method.

The __repr__ method is overridden and returns a string representation of
all menu items for the current day.

the __len__ method is overridden and returns the length of items
for the current day.