New module: Commandprocessor
The new commandprocessor module contains new logic to process speech in chat for the bot.
The commandprocessor is designed to act as a framework for parsing messages from a user
in chat, making it easier to implement advanced features that requires multiple variations
of text to be understood by the bot.

As new features are implemented and build alongside the rest of the code base, the goal
with this new way of parsing text in chat is to mitigate the issues in the current way
of binding keywords to actions. Since some features may use similar keywords but in
different constellations, conflicts could reside and a very primitive way of prioritizing
the features by the order arrangement in the dictionary for matching keywords to actions
was the solution to this.

Again, mitigating this is one of the goals with this new module.