New method: cache_web_content the action to cache a Menu object was moved to this method for cleaner code and the ability for reuse. The caching is the process of scraping the website itself.
New method: get_menu_for_weekday
Get a list of dishes on the menu for a given day of week as
decided by the integer passed(0-5). The implementation uses the
__getitem__ overload in the Menu object.

New method: get_menu_for_week
Get a list with the entire menu for the week. Using the __getitem__
overload in the Menu object by giving it a slice of 0:5.