The CommandProcessor class
The new CommandProcessor class is meant to be instantiated once.
It is the receiving object for the raw command read from the user.
It houses all the Feature instances and holds the algorithm for
decomposing the command in to words, identifying pronouns and matching
it against features with command parsers.

an Interpretation object is returned upon call of the .response() method
which will paint a more complete picture of what information could be
retrieved in the command, as well as the response and in case an error
was encountered, it is also present in the Interpretation object.