Improved subcategory iteration.
With the option for developers to add more complex subcategory keys
to an enum instance that points to a method call, there can be a vast
reduction in the repetition of enums for different constellations of
words and spellings.

Complex subcategories can look like:

{'word': ('any', 'of', 'these', afterwards')}

...Where the word unlocks a tuple, of which the message must then contain
at least one of the elements to comply. It enables more complex and specific
formulations to be required for pointing to a callback in the application.

Simple subcategories as 'this': CommandSubCategory.EXAMPLE  - are still
possible and are used as before.