This file contains two classes:
------- ApiHandle --------

The ApiHandle class is used as a way to interact with an RESTful Api
with caching functionality, as to not query the api more often than
defined upon instantiation.
This class returns data from the api "coronavirus monitor" in this
implementation, but can be used for other api's as well.

-------- Client ----------
The Client class interacts with the dictionary data received by ApiHandle.
There are methods for returning various data points in the response
body from the api.