New Class: PollCache
This class serves as a way to send a method to an object, for checking
whether the return value from this function has changed since the last
time the function was called.
It allows for an endless poll by the caller of the instance of this class,
and only receive data when a difference is detected. The difference is
matched against it exactly, so even the slightest difference in return
value will provide a result.