The PollCache object has been updated to account for individual functions beyond just the function itself that is being called, but to also account for the constellation of arguments and keyword arguments for the given call. These constellations are treated as unique caches and that enables the PollCache instance to keep track and record of many calls with the same function, but with different parameters and not overwrite them just because the function is the same and already stored.
Updated and improved docstring.

silent_first_call implemented to allow for suppression of outputs the
very first time that the methods are called through the PollCache object.
This is useful where the first round of calls are meant to be as a
build up of cache, and the return value isn't of interest the very first
time.