Changes made to the CI framework for version 1.2.0:
* Improvements in the reduction of enums throughout the stack.
Enums are great. But they don't really aid in making it easier to use
the framework, which is the primary goal. Every time a developer would
implement a new feature, or a new set of sub categories or even just the
category, the framework file enumerators.py had to be edited which
felt very unintuitive and nonsensical, since a framework should never
have to be edited by the user.

Therefore, the CommandSubcategory and CommandCategory enums are completely
removed. They are instead replaced with a set of callbacks which replaces
the CommandSubcategory enum, and the CommandCategory (which really only
was used in logging, to see which feature was given the job) is updated
to automatically be shown as the name of the class, that is, the name
of the feature itself, in the log files. This makes it much easier
to develop with CI and the features become a lot less cluttered with
multiple dictionaries.

The get_category method in the CommandParserBase class is renamed
is_contender_for_processing and instead of returning a category, which
actually was parsed as a bool, now in fact returns a bool. It lets the
CommandProcessor object know if the parser matches any criteria that
makes it eligible to work with the command, and try to give a response
for it.

Subcategories property in the FeatureBase is completely removed, since
the CommandParser object takes care of this fully.
It is replaced in the CommandParser to a callbacks property which simply
holds the dictionary with the set of subcategory strings or tuples,
with a binding toward methods from the feature.